+++ docs/changelog.rst
@@ -13,6 +13,7 @@ Changelog
 
 * Windows: fixed wrong Editor View window size.
 * Added a setting to limit the number of lines displayed by the Console.
+* Added hyperlinks to resource paths in the Console.
 
 0.42.0
 ------